Move the docs-header teardown registry into its own module
The teardown added for B-38 pushed docs-header.ts to 379 lines, past its
360 transition cap. I did not run the size gate locally before pushing --
CI caught it.

The registry is a coherent unit on its own, so it moves out rather than
the file being trimmed to fit.
